[FIX] account: fix searchpanel width issue
This commit fixes an issue about the search panel content being cut in
some accounting interfaces. This is due to the fact that the searchpanel
with a specific class would get an arbitrary width value in px.

To avoid this issue, we set a `w-auto` to the searchpanel, allowing it
to adapt its width to its content.

task-3560617

closes odoo/odoo#161382

X-original-commit: f90339c0c6fc27bab7e21d2e78bec043f6327146
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>