[FIX] event: cannot open event if deleted template
Current behavior:
When you delete a mail template that is used
in the communication tab of an event,
you get an error: Missing Record

Steps to reproduce:
1. Go to Email Templates
2. Delete a template used in an event
3. Try to open the event
4. Error: Missing Record

Cause of the issue:
When you delete a template, the fields template_ref
in event.mail and event.type.mail are not changed,
Making reference to nonexistent data.

Fix:
As mentioned in the ticket, lines related
to the template should be unlinked.

opw-3685142

closes odoo/odoo#153893

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>