[FIX] analytic,account: groups are not defined in right module
The fields in analytic have a group which is defined in account.
We hide them in analytic module and show them in account with the right group.
We also show balance now, even if you don't have the group 'readonly',
because it's an information that is independent of credit/debit notion.

Linked to runbot error 59222 (will need to remove the group in the test)

closes odoo/odoo#158994

X-original-commit: 41cded27df79334bcabf55f62441630ed0524118
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>