[FIX] website: fix animated text in translation mode
Steps to reproduce:

- Go to the website translation mode with some animated text in page
content. Be sure to have the "Translation Info" modal displayed (don't
click on the `"Ok, never show me this again"` button).
- The text won't be animated and only remains invisible.

A fix was introduced in [1] to get the right `scrollTop` value (needed
to handle the change in position from animations) when a modal is opened
in the editable area. In translation mode, the "Translation Info" modal
will just break this check, since it will be mounted outside the
targeted DOM.

The goal of this commit is just to make sure the code won't get this
value from a modal outside the targeted area.

[1]: https://github.com/odoo/odoo/commit/48bb9efee99d3bee2b4b60eec0f59ed2cb13a096

opw-3686777

Part-of: odoo/odoo#157087
Co-authored-by: qsm-odoo <qsm@odoo.com>