[FIX] l10n_ch: tax report box 200 computation fix
With this PR https://github.com/odoo/odoo/pull/129717 we fixed
multiple problems of the Swiss tax report but the 200 box is still
not correct.

Fix the computation of the box 200 of the Swiss tax report.
We remove 382 and 383 boxes that are for purchases and shouldn't be
included.
We add box 205 and 289 so that the box 200 somehow represent the "gross"
taxed amount, and 299 is therefore the "net" amount.

200 = 302 to 343 + 205 + 289

opw-3766215

closes odoo/odoo#158406

X-original-commit: 0042b9d3eece219d4e89fe9c6ecb9971a1f3bf12
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>