[IMP] account: fix reversal invoice date
Reversing an invoice wizard provides an option to use the original entry date, but does not set the invoice_date to the same date.
This behaviour is strange because the UI displays today's date, but the move is actually posted on a different date.

This commit, keeps the dates in sync.
see PR for the source of this fix.

closes odoo/odoo#158640

X-original-commit: dcc6c3b982d2caf6761b3cf45b90406ee4696a7d
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>