[FIX] purchase_stock: 'propagate_cancel' check not visible
Activate "Reception Report" feature
Create a SO for a storable product, confirm.
Create a PO for the same product.
Confirm the PO and check the delivery, open the "Allocation" report
Assign the Product to the delivery of the SO.
Go back to the PO and cancel the order, delivery of the SO will be cancelled.

Issue: Currently the user cannot modify this behavior as the
`propagate_cancel` checkbox is unaccessible

opw-3733512

closes odoo/odoo#158587

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>