[REM] base: broken method `Partner._email_send`
The underlying `odoo.tools.email_send` was removed in
82de620424e22f9ca7cb9768db638c3a2b4f9318 (merged in 14.5) but this
callsite was missed, this method has been broken ever since.

I really want to remove this method directly, but technically it's an
expensive no-op if called on a recordset of partners without emails
set (or an empty recordset), so instead make it trigger a warning &
remove in master.

closes odoo/odoo#159113

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>