[FIX] base: change the adress format in Luxembourg
Steps to reproduce:

Create and print an SO for a customer based in Luxembourg

Expected behavior:

According to Bpost and to the Post of Luxembourg, the zip code should be
displayed before the city name in the address format in Luxembourg.

Current behavior:

The zip code is displayed after the city name.

opw-3791142

closes odoo/odoo#159240

Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>