[FIX] l10n_id_efaktur: used partner.
At the moment, the partner being used to get the information
for the efaktur module is the partner set on the invoice.

This is wrong, since if the partner set on the invoice is a
child record of a company partner you cannot set most of these
fields.

To fix this issue, we will update in order to fetch the efaktur
related fields from the commercial partner instead. (vat, along
with all efaktur specific information used in the process)

closes odoo/odoo#160108

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>