[PERF] analytic: add missing indexes
Description:
Add indexes affecting queries related to `account.analytic.line`.
Based on heuristics like domains in `search`/`read_group` that are
not really selective, or being an inverse of One2many.
Some of these searches are on code-paths that are frequent, like
creation of an SO.
The goal is to reduce the number of Seq.Scan or non-selective Index
Scan.

Reference:
task-3724844

closes odoo/odoo#160472

Signed-off-by: Olivier Dony (odo) <odo@odoo.com>