[FIX] website: preserve theme images order
The image displayed in the theme selection page is the last one that
was created out of the theme's manifest's `images` entry.
Because of this, when a theme has a different non-last image in a new
version compared to an old version, that image becomes the last created
one and is displayed in the theme selection page instead of the expected
screenshot image.

This commit preserves the order that was specified in the `images`
definition by re-creating all images instead of only the modified ones -
if an update is needed at all.

Steps to reproduce:
- Install `website` with a theme in 14.0.
- Upgrade to 15.0.
- Go to the theme selection page.

=> Some themes (e.g. `theme_beauty`) displayed their cover picture
instead of their screenshot.

task-2719425

closes odoo/odoo#160452

Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>