[FIX] web_editor: remove useless zws even next to the selection
The sanitizer removes ZWSP characters (unicode 200B) added by
`fillEmpty` (parent flagged with the `oe-zws-empty-inline` attribute)
when their parent actually has content. It did so only when the
selection wasn't in said parent, to avoid breaking the selection. But
this means the content is actually wrong in these cases. This commit
removes these ZWSP characters anyway, and corrects the selection if
needed.

Part-of: odoo/odoo#160718