[FIX] pos : product images look zoomed in
The cover scss rule first stretches the image to fill the container completely,
then cropped at the size of the container.

This results in some poor display result if the image has a weird aspect ratio.

This is a behaviour change from saas-16.3 where the image wasn't cropped and
simply resized to fit inside the container.

opw-3826349

closes odoo/odoo#161111

X-original-commit: 0cc3a0edfcd68ebbd7930eedd2af82bd3b3a29fe
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Loïc Leloup (lole) <lole@odoo.com>