[FIX] analytic: plan's applicability don't work without companies
If you create an applicability and remove the company field,
they are never used.
An applicability like this should be valid for all companies.
We put a 0.5 value for the company field so an applicability
so it has a lesser priority than other fields.
Same idea as the distribution models.

opw-3847415

closes odoo/odoo#162152

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>