[REF] phone_validation: phonenumbers monkey patches redesigned approach
I Test purpose
Before tests were doing 2 things:
- checking if monkey patch was applied (with rather complicated check)
- checking if numbers that we consider correct are correct
After
- tests are only responsible for ensuring that numbers are passed corrctly

II was patch applied?
    What about checking if patch was applied? Well that depends strictly
on the version of phonenumbers that got installed. So if the version of
phonenumber is withing boundaries defined for patching then we patch
else we don't patch.
    Logging mechanism was added such that we'll know that we started
using newer version of phonenumbers and the monkeypatch can be removed
from the code

III Test consistency
    Before each test had slightly different approach to checking whether
parsing is performed correctly, but they were doing so with different
granuality: some tests were only checking whether number is valid, some
if parsed country code is correct
    After the refactor we have one-for-all testing approach which
is based on creating a dataset of inputs and outputs. We run parsing for
all the inputs and then we perform checks against given outputs (which are
considered to be ground-truths). We can specyfy all the outputs (we
check region and national number of the parsed results), some of them, or
none (in which case we'd only check simple validity of parsed number).

IV Naming convetions
coutry code in difrent sources can mean either literal ISO code (eg. "SN")
or numeral code (eg. +221). This commit sticks to the following
convention
"region" always refers to the 2 letter literal ISO country code (eg. "SN")
"code" - always refers to number code (eg. +221)

closes odoo/odoo#156442

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>