[FIX] website_sale: fix tour due to l10n_ar
Backport of https://github.com/odoo/odoo/commit/cfbe95ae05fd0c87e24668fc8550da9fdd7ee70d to solve
issue in `google_analytics_add_to_cart` due to renaming of the product.

At the same time, some tours were broken when installed l10n_ar_website_sale
`shop_buy_product` fails because there are missing information on the
company/partner.
Fixed it from the python, as the override from the js is complex
(as the tour is already overriden in a module that is not in his dependencies).

Linked to runbot errors 6169,6170,6172

closes odoo/odoo#163485

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>