[FIX] account: backport more rounding under round globally for EDI
Currently the tax amounts returned by `_prepare_edi_tax_details`
are not rounded.
But in the UI / DB the tax values are automatically rounded since
they are put on account move lines.

But i.e. the rounded sum of unrounded values is not necessarily the same as
the sum of the rounded values:
e.g. (c.f. invoice `test_invoice_withholded_taxes` from module `l10n_co_edi`)
    2.87565 + 0.7575 ≈ 3.6331499999999997 (rounds to 3.63)
    but
    2.88 + 0.76 = 3.64

In 16.0+ there is already some logic in place s.t. the round globally tax values
are rounded.
(see i.e. commit deffd01163e6872c304fcd0d01dfa472ba95f2d6)
This commit backports the current state of this logic to 15.

Compared to 16.0 the computation of the `grouping_key` had to be changed.
It was done like in function `_recompute_tax_lines` of model 'account.move'.

Part-of: odoo/odoo#148953